STILLWORKS-BACKEND/requirements.txt
Flask==3.1.0
Flask-PyMongo==2.3.0
Flask-CORS==5.0.0
Flask-JWT-Extended==4.7.1
Flask-Bcrypt==1.0.1
pymongo[srv]==4.10.1
python-dotenv==1.0.1
Werkzeug==3.1.3
Pillow==11.1.0
gunicorn==23.0.0
flask-jwt-extended==4.5.2
